testing: move test_settings fixture to unit/conftest
